Fix unnecesery removing
The error was next:
If in one day the strategy has return, but the benchmark has no data, df will have next view (for example): 0.1 na
The old method removes the return of strategy. 
In result, if we compare the strategy to different benchmarks, we will result to different numbers in strategy return. Its frustrating.