Added Theming
This is a rather large change in terms of the amount of code affected owing
to it proving rather difficult to add this feature into things in a reasonable
manner. The following changes have been made:
* collect all plotting functions in a single class, where `theme` is defined at instantiation
* all uses of the `grayscale` keyword have been either replaced with `theme` or removed as appropriate
* the `options` keyword has been removed from `qs.html()`
* as well, the `download_filename` keyword has been replaces with `html_filename`
* type hinting was added to all code touched
* all non-conventional import statements have been made conventional
* the 'Known Issues' segment of the `README` has been removed as the issue is no longer present

To use theming, a the current `default` and `grayscale` themes have been provided
in `_plotting/themes.py`. Any user-defined theme can be be created by following the
dictionary structure provided, noting that all keys must be present or the behaviour
is to fallback to `default`